Move ambiguity detection code into its own file
